pyproject.toml
[project]
name = "namestamp"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
  "colorama>=0.4.0",
  "pillow>=12.0.0",
  "pyinstaller>=6.17.0",
  "pywin32>=311",
]
